move Mapper to training.py
